# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API do Sistema Financeiro
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categorias:
    get:
      operationId: listCategorias
      tags: [categorias]
      summary: Listar categorias
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
            enum: [receita, despesa, todas]
      responses:
        "200":
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Categoria"
    post:
      operationId: createCategoria
      tags: [categorias]
      summary: Criar categoria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoriaInput"
      responses:
        "201":
          description: Categoria criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categoria"
  /categorias/{id}:
    get:
      operationId: getCategoria
      tags: [categorias]
      summary: Buscar categoria por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Categoria
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categoria"
        "404":
          description: Não encontrado
    patch:
      operationId: updateCategoria
      tags: [categorias]
      summary: Atualizar categoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoriaUpdate"
      responses:
        "200":
          description: Categoria atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Categoria"
    delete:
      operationId: deleteCategoria
      tags: [categorias]
      summary: Excluir categoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Categoria excluída

  # ─────────────────────────── CONTAS ───────────────────────────
  /contas:
    get:
      operationId: listContas
      tags: [contas]
      summary: Listar contas
      responses:
        "200":
          description: Lista de contas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conta"
    post:
      operationId: createConta
      tags: [contas]
      summary: Criar conta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContaInput"
      responses:
        "201":
          description: Conta criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conta"
  /contas/{id}:
    get:
      operationId: getConta
      tags: [contas]
      summary: Buscar conta por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conta"
    patch:
      operationId: updateConta
      tags: [contas]
      summary: Atualizar conta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContaUpdate"
      responses:
        "200":
          description: Conta atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conta"
    delete:
      operationId: deleteConta
      tags: [contas]
      summary: Excluir conta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Conta excluída

  # ─────────────────────────── TRANSAÇÕES ───────────────────────────
...
  /contas-a-pagar:
    get:
      operationId: listContasAPagar
      tags: [contasAPagar]
      summary: Listar contas a pagar
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, pago, vencido, cancelado]
        - name: dataInicio
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: dataFim
          in: query
          schema:
            type: ["string", "null"]
            format: date
        - name: busca
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Lista de contas a pagar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAPagarListResponse"
    post:
      operationId: createContaAPagar
      tags: [contasAPagar]
      summary: Criar conta a pagar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContaAPagarInput"
      responses:
        "201":
          description: Conta a pagar criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAPagar"
  /contas-a-pagar/{id}:
    get:
      operationId: getContaAPagar
      tags: [contasAPagar]
      summary: Buscar conta a pagar por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conta a pagar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAPagar"
        "404":
          description: Não encontrado
    patch:
      operationId: updateContaAPagar
      tags: [contasAPagar]
      summary: Atualizar conta a pagar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContaAPagarUpdate"
      responses:
        "200":
          description: Conta a pagar atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAPagar"
    delete:
      operationId: deleteContaAPagar
      tags: [contasAPagar]
      summary: Excluir conta a pagar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Excluído
  /contas-a-pagar/{id}/pagar:
    post:
      operationId: pagarContaAPagar
      tags: [contasAPagar]
      summary: Marcar conta a pagar como paga
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagamentoInput"
      responses:
        "200":
          description: Conta paga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContaAPagar"

  # ─────────────────────────── CONTAS A RECEBER ───────────────────────────
...
components:
  schemas:
...
    Categoria:
      type: object
      required: [id, nome, tipo, cor, icone, createdAt]
      properties:
        id:
          type: integer
        nome:
          type: string
        tipo:
          type: string
          enum: [receita, despesa]
        cor:
          type: string
        icone:
          type: string
        descricao:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CategoriaInput:
      type: object
      required: [nome, tipo, cor, icone]
      properties:
        nome:
          type: string
          minLength: 1
        tipo:
          type: string
          enum: [receita, despesa]
        cor:
          type: string
        icone:
          type: string
        descricao:
          type: string
    CategoriaUpdate:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        tipo:
          type: string
          enum: [receita, despesa]
        cor:
          type: string
        icone:
          type: string
        descricao:
          type: string

    # ─────── CONTAS ───────
    Conta:
      type: object
      required: [id, nome, tipo, saldoInicial, saldoAtual, ativa, createdAt]
      properties:
        id:
          type: integer
        nome:
          type: string
        tipo:
          type: string
          enum: [corrente, poupanca, investimento, carteira, outro]
        banco:
          type: ["string", "null"]
        saldoInicial:
          type: number
        saldoAtual:
          type: number
        cor:
          type: ["string", "null"]
        ativa:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ContaInput:
      type: object
      required: [nome, tipo, saldoInicial]
      properties:
        nome:
          type: string
          minLength: 1
        tipo:
          type: string
          enum: [corrente, poupanca, investimento, carteira, outro]
        banco:
          type: string
        saldoInicial:
          type: number
        cor:
          type: string
    ContaUpdate:
      type: object
      properties:
        nome:
          type: string
        tipo:
          type: string
          enum: [corrente, poupanca, investimento, carteira, outro]
        banco:
          type: string
        cor:
          type: string
        ativa:
          type: boolean

    # ─────── TRANSAÇÕES ───────
...
    ContaAPagar:
      type: object
      required: [id, descricao, valor, dataVencimento, status, createdAt]
      properties:
        id:
          type: integer
        descricao:
          type: string
        fornecedor:
          type: ["string", "null"]
        valor:
          type: number
        dataVencimento:
          type: string
          format: date
        dataPagamento:
          type: ["string", "null"]
          format: date
        valorPago:
          type: ["number", "null"]
        status:
          type: string
          enum: [pendente, pago, vencido, cancelado]
        observacoes:
          type: ["string", "null"]
        categoriaId:
          type: ["integer", "null"]
        categoriaNome:
          type: ["string", "null"]
        contaId:
          type: ["integer", "null"]
        contaNome:
          type: ["string", "null"]
        recorrente:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ContaAPagarInput:
      type: object
      required: [descricao, valor, dataVencimento]
      properties:
        descricao:
          type: string
          minLength: 1
        fornecedor:
          type: string
        valor:
          type: number
          minimum: 0.01
        dataVencimento:
          type: string
          format: date
        observacoes:
          type: string
        categoriaId:
          type: ["integer", "null"]
        contaId:
          type: ["integer", "null"]
        recorrente:
          type: boolean
    ContaAPagarUpdate:
      type: object
      properties:
        descricao:
          type: string
        fornecedor:
          type: string
        valor:
          type: number
        dataVencimento:
          type: string
          format: date
        status:
          type: string
          enum: [pendente, pago, vencido, cancelado]
        observacoes:
          type: string
        categoriaId:
          type: ["integer", "null"]
        contaId:
          type: ["integer", "null"]
        recorrente:
          type: boolean
    PagamentoInput:
      type: object
      required: [dataPagamento]
      properties:
        dataPagamento:
          type: string
          format: date
        valorPago:
          type: number
        contaId:
          type: ["integer", "null"]
    ContaAPagarListResponse:
      type: object
      required: [data, total, page, limit, totalPendente, totalVencido]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ContaAPagar"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPendente:
          type: number
        totalVencido:
          type: number

    # ─────── CONTAS A RECEBER ───────
...