# lib/api-zod/src/generated/types/categoriaInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API do Sistema Financeiro
 * OpenAPI spec version: 0.1.0
 */
import type { CategoriaInputTipo } from './categoriaInputTipo';

export interface CategoriaInput {
  /** @minLength 1 */
  nome: string;
  tipo: CategoriaInputTipo;
  cor: string;
  icone: string;
  descricao?: string;
}