# lib/api-zod/src/generated/types/contaAPagarInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API do Sistema Financeiro
 * OpenAPI spec version: 0.1.0
 */

export interface ContaAPagarInput {
  /** @minLength 1 */
  descricao: string;
  fornecedor?: string;
  /** @minimum 0.01 */
  valor: number;
  dataVencimento: Date;
  observacoes?: string;
  /** @nullable */
  categoriaId?: number | null;
  /** @nullable */
  contaId?: number | null;
  recorrente?: boolean;
}