# lib/api-zod/src/generated/types/contaAPagarUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API do Sistema Financeiro
 * OpenAPI spec version: 0.1.0
 */
import type { ContaAPagarUpdateStatus } from './contaAPagarUpdateStatus';

export interface ContaAPagarUpdate {
  descricao?: string;
  fornecedor?: string;
  valor?: number;
  dataVencimento?: Date;
  status?: ContaAPagarUpdateStatus;
  observacoes?: string;
  /** @nullable */
  categoriaId?: number | null;
  /** @nullable */
  contaId?: number | null;
  recorrente?: boolean;
}