# lib/api-zod/src/generated/types/contaInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API do Sistema Financeiro
 * OpenAPI spec version: 0.1.0
 */
import type { ContaInputTipo } from './contaInputTipo';

export interface ContaInput {
  /** @minLength 1 */
  nome: string;
  tipo: ContaInputTipo;
  banco?: string;
  saldoInicial: number;
  cor?: string;
}